Add files via upload
I trim the first 40 and last 40 data points of the sensitivity function, so now all the actual data is conserved. 
I do delete a couple of pixels after the flux calibration, but this can be done in either the chip gap seperation step or before the flux calibration. I plan to make it before the flux clalibration,  choosing the trimming sections interactively as if one would do it in pyraf.